@@ -35,7 +35,7 @@ import android.preference.Preference;
 import android.preference.PreferenceCategory;
 import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.widget.Toolbar;
 import android.view.ViewStub;
 import android.widget.Toast;
 
@@ -53,6 +53,7 @@ import org.shadowice.flocke.andotp.Utilities.Settings;
 import org.shadowice.flocke.andotp.Utilities.UIHelper;
 
 import java.util.ArrayList;
+import java.util.Locale;
 
 import javax.crypto.SecretKey;
 
@@ -148,7 +149,7 @@ public class SettingsActivity extends BaseActivity
         if (fragment.useAutoBackup != null) {
             fragment.useAutoBackup.setEnabled(BackupHelper.autoBackupType(this) == Constants.BackupType.ENCRYPTED);
             if (!fragment.useAutoBackup.isEnabled())
-                fragment.useAutoBackup.setChecked(false);
+                fragment.useAutoBackup.setValue(Constants.AutoBackup.OFF.toString().toLowerCase(Locale.ENGLISH));
         }
     }
 
@@ -243,7 +244,7 @@ public class SettingsActivity extends BaseActivity
 
         Settings settings;
         ListPreference encryption;
-        CheckBoxPreference useAutoBackup;
+        ListPreference useAutoBackup;
         CheckBoxPreference useAndroidSync;
 
         OpenPgpAppPreference pgpProvider;
@@ -352,10 +353,10 @@ public class SettingsActivity extends BaseActivity
                 }
             });
 
-            useAutoBackup = (CheckBoxPreference) findPreference(getString(R.string.settings_key_auto_backup_password_enc));
+            useAutoBackup = (ListPreference)findPreference(getString(R.string.settings_key_auto_backup_password_enc));
             useAutoBackup.setEnabled(BackupHelper.autoBackupType(getActivity()) == Constants.BackupType.ENCRYPTED);
             if(!useAutoBackup.isEnabled())
-                useAutoBackup.setChecked(false);
+                useAutoBackup.setValue(Constants.AutoBackup.OFF.toString().toLowerCase(Locale.ENGLISH));
 
             useAndroidSync = (CheckBoxPreference) findPreference(getString(R.string.settings_key_enable_android_backup_service));
             useAndroidSync.setEnabled(settings.getEncryption() == EncryptionType.PASSWORD);
