@@ -28,7 +28,7 @@ import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.Canvas;
 import android.graphics.drawable.Drawable;
-import android.support.v7.app.AppCompatDelegate;
+import androidx.appcompat.app.AppCompatDelegate;
 
 import org.shadowice.flocke.andotp.R;
 
@@ -40,6 +40,7 @@ public class EntryThumbnail {
 
     public enum EntryThumbnails {
         Default(R.mipmap.ic_launcher_round),
+        OneAndOne(R.drawable.thumb_1and1),
         OnePassword(R.drawable.thumb_1password),
         Adobe(R.drawable.thumb_adobe),
         AdGuard(R.drawable.thumb_adguard),
@@ -61,6 +62,7 @@ public class EntryThumbnail {
         BitBucket(R.drawable.thumb_bitbucket),
         Bitcoin(R.drawable.thumb_bitcoin),
         Bitfinex(R.drawable.thumb_bitfinex),
+        Bitmex(R.drawable.thumb_bitmex),
         Bitstamp(R.mipmap.thumb_bitstamp, AssetType.Bitmap),
         Bitpanda(R.drawable.thumb_bitpanda),
         Bittrex(R.drawable.thumb_bittrex),
@@ -86,6 +88,7 @@ public class EntryThumbnail {
         Discord(R.drawable.thumb_discord),
         Discourse(R.drawable.thumb_discourse),
         Disroot(R.drawable.thumb_disroot),
+        Docker(R.drawable.thumb_docker),
         DocuSign(R.drawable.thumb_docusign),
         Dropbox(R.drawable.thumb_dropbox),
         ElectronicArts(R.drawable.thumb_electronic_arts),
@@ -95,6 +98,7 @@ public class EntryThumbnail {
         Facebook(R.drawable.thumb_facebook),
         FACEIT(R.drawable.thumb_faceit),
         Fingerprint(R.drawable.thumb_fingerprint),
+        Finnair(R.drawable.thumb_finnair),
         Firefox(R.drawable.thumb_firefox),
         Flight(R.drawable.thumb_flight_takeoff),
         Floatplane(R.drawable.thumb_floatplane),
@@ -116,6 +120,7 @@ public class EntryThumbnail {
         HomeAssistant(R.drawable.thumb_home_assistant),
         HumbleBundle(R.drawable.thumb_humblebundle),
         HurricaneElectric(R.drawable.thumb_hurricane_electric),
+        IBM(R.drawable.thumb_ibm),
         Iconomi(R.drawable.thumb_iconomi),
         IFTTT(R.drawable.thumb_ifttt),
         Instagram(R.drawable.thumb_instagram),
