@@ -29,9 +29,9 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.net.Uri;
 import android.os.Handler;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.PopupMenu;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.appcompat.widget.PopupMenu;
+import androidx.recyclerview.widget.RecyclerView;
 import android.text.Editable;
 import android.text.InputType;
 import android.text.TextWatcher;
@@ -50,6 +50,7 @@ import android.widget.LinearLayout;
 import android.widget.Toast;
 
 import org.shadowice.flocke.andotp.Database.Entry;
+import org.shadowice.flocke.andotp.Dialogs.TagsDialog;
 import org.shadowice.flocke.andotp.R;
 import org.shadowice.flocke.andotp.Utilities.BackupHelper;
 import org.shadowice.flocke.andotp.Utilities.Constants;
@@ -119,7 +120,7 @@ public class EntriesCardAdapter extends RecyclerView.Adapter<EntryViewHolder>
         if (! entries.contains(e)) {
             entries.add(e);
             entriesChanged();
-            saveEntries(true);
+            saveEntries(settings.getAutoBackupEncryptedPasswordsEnabled());
         } else {
             Toast.makeText(context, R.string.toast_entry_exists, Toast.LENGTH_LONG).show();
         }
@@ -138,7 +139,7 @@ public class EntriesCardAdapter extends RecyclerView.Adapter<EntryViewHolder>
     public void saveEntries(boolean auto_backup) {
         DatabaseHelper.saveDatabase(context, entries, encryptionKey);
 
-        if(auto_backup && settings.getAutoBackupEncryptedPasswordsEnabled()) {
+        if(auto_backup) {
             Constants.BackupType backupType = BackupHelper.autoBackupType(context);
             if (backupType == Constants.BackupType.ENCRYPTED) {
                 Uri backupFilename = Tools.buildUri(settings.getBackupDir(), BackupHelper.backupFilename(context, Constants.BackupType.ENCRYPTED));
@@ -321,7 +322,7 @@ public class EntriesCardAdapter extends RecyclerView.Adapter<EntryViewHolder>
         realEntry.setCounter(counter);
         realEntry.updateOTP();
         
-        saveEntries(false);
+        saveEntries(settings.getAutoBackupEncryptedFullEnabled());
     }
 
     private void hideEntry(Entry entry) {
@@ -374,7 +375,7 @@ public class EntriesCardAdapter extends RecyclerView.Adapter<EntryViewHolder>
                         Entry e = entries.get(realIndex);
                         e.setCounter(newCounter);
 
-                        saveEntries(false);
+                        saveEntries(settings.getAutoBackupEncryptedFullEnabled());
                     }
                 })
                 .setNegativeButton(android.R.string.cancel, new DialogInterface.OnClickListener() {
@@ -392,7 +393,7 @@ public class EntriesCardAdapter extends RecyclerView.Adapter<EntryViewHolder>
             displayedEntries.get(position).setLastUsed(timeStamp);
 
         entries.get(realIndex).setLastUsed(timeStamp);
-        saveEntries(false);
+        saveEntries(settings.getAutoBackupEncryptedFullEnabled());
 
         if (sortMode == SortMode.LAST_USED) {
             displayedEntries = sortEntries(displayedEntries);
@@ -411,7 +412,7 @@ public class EntriesCardAdapter extends RecyclerView.Adapter<EntryViewHolder>
             displayedEntries = new ArrayList<>(entries);
             notifyItemMoved(fromPosition, toPosition);
 
-            saveEntries(false);
+            saveEntries(settings.getAutoBackupEncryptedFullEnabled());
         }
 
         return true;
@@ -496,7 +497,7 @@ public class EntriesCardAdapter extends RecyclerView.Adapter<EntryViewHolder>
                         Entry e = entries.get(realIndex);
                         e.setLabel(newLabel);
 
-                        saveEntries(false);
+                        saveEntries(settings.getAutoBackupEncryptedFullEnabled());
                     }
                 })
                 .setNegativeButton(android.R.string.cancel, new DialogInterface.OnClickListener() {
@@ -582,7 +583,7 @@ public class EntriesCardAdapter extends RecyclerView.Adapter<EntryViewHolder>
                 Entry e = entries.get(realIndex);
                 e.setThumbnail(thumbnail);
 
-                saveEntries(false);
+                saveEntries(settings.getAutoBackupEncryptedFullEnabled());
                 notifyItemChanged(pos);
                 alert.cancel();
             }
@@ -610,7 +611,7 @@ public class EntriesCardAdapter extends RecyclerView.Adapter<EntryViewHolder>
             @Override
             public Object call() throws Exception {
                 entries.get(realPos).setTags(tagsAdapter.getActiveTags());
-                saveEntries(false);
+                saveEntries(settings.getAutoBackupEncryptedFullEnabled());
 
                 List<String> inUseTags = getTags();
 
@@ -655,7 +656,7 @@ public class EntriesCardAdapter extends RecyclerView.Adapter<EntryViewHolder>
                         notifyItemRemoved(pos);
 
                         entries.remove(realIndex);
-                        saveEntries(false);
+                        saveEntries(settings.getAutoBackupEncryptedFullEnabled());
                     }
                 })
                 .setNegativeButton(android.R.string.no, new DialogInterface.OnClickListener() {
